% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouderrorreporting_functions.R
\docType{package}
\name{clouderrorreporting_googleAuthR}
\alias{clouderrorreporting_googleAuthR}
\alias{clouderrorreporting_googleAuthR-package}
\title{Stackdriver Error Reporting API
Stackdriver Error Reporting groups and counts similar errors from cloud services. The Stackdriver Error Reporting API provides a way to report new errors and read access to error groups and their associated errors.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 19:32:10
filename: /Users/mark/dev/R/autoGoogleAPI/googleclouderrorreportingv1beta1.auto/R/clouderrorreporting_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}
}
