% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouderrorreporting_objects.R
\name{DeleteEventsResponse}
\alias{DeleteEventsResponse}
\title{DeleteEventsResponse Object}
\usage{
DeleteEventsResponse()
}
\value{
DeleteEventsResponse object
}
\description{
DeleteEventsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for deleting error events.
}
\seealso{
Other res: \code{\link{ErrorContext}},
  \code{\link{ErrorEvent}}, \code{\link{ErrorGroupStats}},
  \code{\link{ErrorGroup}},
  \code{\link{HttpRequestContext}},
  \code{\link{ListEventsResponse}},
  \code{\link{ListGroupStatsResponse}},
  \code{\link{ReportErrorEventResponse}},
  \code{\link{ReportedErrorEvent}},
  \code{\link{ServiceContext}},
  \code{\link{SourceLocation}}, \code{\link{TimedCount}},
  \code{\link{TrackingIssue}}
}

