% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouderrorreporting_objects.R
\name{ErrorContext}
\alias{ErrorContext}
\title{ErrorContext Object}
\usage{
ErrorContext(httpRequest = NULL, reportLocation = NULL, user = NULL)
}
\arguments{
\item{httpRequest}{The HTTP request which was processed when the error was}

\item{reportLocation}{The location in the source code where the decision was made to}

\item{user}{The user who caused or was affected by the crash}
}
\value{
ErrorContext object
}
\description{
ErrorContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A description of the context in which an error occurred.This data should be provided by the application when reporting an error,unless theerror report has been generated automatically from Google App Engine logs.
}
\seealso{
Other res: \code{\link{DeleteEventsResponse}},
  \code{\link{ErrorEvent}}, \code{\link{ErrorGroupStats}},
  \code{\link{ErrorGroup}},
  \code{\link{HttpRequestContext}},
  \code{\link{ListEventsResponse}},
  \code{\link{ListGroupStatsResponse}},
  \code{\link{ReportErrorEventResponse}},
  \code{\link{ReportedErrorEvent}},
  \code{\link{ServiceContext}},
  \code{\link{SourceLocation}}, \code{\link{TimedCount}},
  \code{\link{TrackingIssue}}
}

