% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouderrorreporting_objects.R
\name{ErrorEvent}
\alias{ErrorEvent}
\title{ErrorEvent Object}
\usage{
ErrorEvent(serviceContext = NULL, eventTime = NULL, context = NULL,
  message = NULL)
}
\arguments{
\item{serviceContext}{The `ServiceContext` for which this error was reported}

\item{eventTime}{Time when the event occurred as provided in the error report}

\item{context}{Data about the context in which the error occurred}

\item{message}{The stack trace that was reported or logged by the service}
}
\value{
ErrorEvent object
}
\description{
ErrorEvent Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An error event which is returned by the Error Reporting system.
}
