% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouderrorreporting_objects.R
\name{ErrorGroup}
\alias{ErrorGroup}
\title{ErrorGroup Object}
\usage{
ErrorGroup(groupId = NULL, name = NULL, trackingIssues = NULL)
}
\arguments{
\item{groupId}{Group IDs are unique for a given project}

\item{name}{The group resource name}

\item{trackingIssues}{Associated tracking issues}
}
\value{
ErrorGroup object
}
\description{
ErrorGroup Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Description of a group of similar error events.
}
\seealso{
Other res: \code{\link{DeleteEventsResponse}},
  \code{\link{ErrorContext}}, \code{\link{ErrorEvent}},
  \code{\link{ErrorGroupStats}},
  \code{\link{HttpRequestContext}},
  \code{\link{ListEventsResponse}},
  \code{\link{ListGroupStatsResponse}},
  \code{\link{ReportErrorEventResponse}},
  \code{\link{ReportedErrorEvent}},
  \code{\link{ServiceContext}},
  \code{\link{SourceLocation}}, \code{\link{TimedCount}},
  \code{\link{TrackingIssue}}
}

