% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouderrorreporting_objects.R
\name{ErrorGroupStats}
\alias{ErrorGroupStats}
\title{ErrorGroupStats Object}
\usage{
ErrorGroupStats(representative = NULL, numAffectedServices = NULL,
  affectedUsersCount = NULL, count = NULL, firstSeenTime = NULL,
  lastSeenTime = NULL, group = NULL, affectedServices = NULL,
  timedCounts = NULL)
}
\arguments{
\item{representative}{An arbitrary event that is chosen as representative for the whole group}

\item{numAffectedServices}{The total number of services with a non-zero error count for the given}

\item{affectedUsersCount}{Approximate number of affected users in the given group that}

\item{count}{Approximate total number of events in the given group that match}

\item{firstSeenTime}{Approximate first occurrence that was seen for this group and}

\item{lastSeenTime}{Approximate last occurrence that was seen for this group}

\item{group}{Group data that is independent of the selection criteria}

\item{affectedServices}{Service contexts with a non-zero error count for the given selection}

\item{timedCounts}{Approximate number of occurrences over time}
}
\value{
ErrorGroupStats object
}
\description{
ErrorGroupStats Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Data extracted for a specific group based on certain selection criteria,such as a given time period and/or service filter.
}
\seealso{
Other res: \code{\link{DeleteEventsResponse}},
  \code{\link{ErrorContext}}, \code{\link{ErrorEvent}},
  \code{\link{ErrorGroup}},
  \code{\link{HttpRequestContext}},
  \code{\link{ListEventsResponse}},
  \code{\link{ListGroupStatsResponse}},
  \code{\link{ReportErrorEventResponse}},
  \code{\link{ReportedErrorEvent}},
  \code{\link{ServiceContext}},
  \code{\link{SourceLocation}}, \code{\link{TimedCount}},
  \code{\link{TrackingIssue}}
}

