% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouderrorreporting_objects.R
\name{HttpRequestContext}
\alias{HttpRequestContext}
\title{HttpRequestContext Object}
\usage{
HttpRequestContext(method = NULL, responseStatusCode = NULL,
  remoteIp = NULL, url = NULL, referrer = NULL, userAgent = NULL)
}
\arguments{
\item{method}{The type of HTTP request, such as `GET`, `POST`, etc}

\item{responseStatusCode}{The HTTP response status code for the request}

\item{remoteIp}{The IP address from which the request originated}

\item{url}{The URL of the request}

\item{referrer}{The referrer information that is provided with the request}

\item{userAgent}{The user agent information that is provided with the request}
}
\value{
HttpRequestContext object
}
\description{
HttpRequestContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
HTTP request data that is related to a reported error.This data should be provided by the application when reporting an error,unless theerror report has been generated automatically from Google App Engine logs.
}
\seealso{
Other res: \code{\link{DeleteEventsResponse}},
  \code{\link{ErrorContext}}, \code{\link{ErrorEvent}},
  \code{\link{ErrorGroupStats}}, \code{\link{ErrorGroup}},
  \code{\link{ListEventsResponse}},
  \code{\link{ListGroupStatsResponse}},
  \code{\link{ReportErrorEventResponse}},
  \code{\link{ReportedErrorEvent}},
  \code{\link{ServiceContext}},
  \code{\link{SourceLocation}}, \code{\link{TimedCount}},
  \code{\link{TrackingIssue}}
}

