% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouderrorreporting_objects.R
\name{ListEventsResponse}
\alias{ListEventsResponse}
\title{ListEventsResponse Object}
\usage{
ListEventsResponse(timeRangeBegin = NULL, errorEvents = NULL,
  nextPageToken = NULL)
}
\arguments{
\item{timeRangeBegin}{The timestamp specifies the start time to which the request was restricted}

\item{errorEvents}{The error events which match the given request}

\item{nextPageToken}{If non-empty, more results are available}
}
\value{
ListEventsResponse object
}
\description{
ListEventsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains a set of requested error events.
}
