% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouderrorreporting_functions.R
\name{projects.deleteEvents}
\alias{projects.deleteEvents}
\title{Deletes all error events of a given project.}
\usage{
projects.deleteEvents(projectName)
}
\arguments{
\item{projectName}{[Required] The resource name of the Google Cloud Platform project}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/error-reporting/}{Google Documentation}
}

