% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouderrorreporting_objects.R
\name{ReportedErrorEvent}
\alias{ReportedErrorEvent}
\title{Stackdriver Error Reporting API Objects 
Stackdriver Error Reporting groups and counts similar errors from cloud services. The Stackdriver Error Reporting API provides a way to report new errors and read access to error groups and their associated errors.}
\usage{
ReportedErrorEvent(serviceContext = NULL, context = NULL,
  eventTime = NULL, message = NULL)
}
\arguments{
\item{serviceContext}{[Required] The service context in which this error has occurred}

\item{context}{[Optional] A description of the context in which the error occurred}

\item{eventTime}{[Optional] Time when the event occurred}

\item{message}{[Required] A message describing the error}
}
\value{
ReportedErrorEvent object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 22:59:35
filename: /Users/mark/dev/R/autoGoogleAPI/googleclouderrorreportingv1beta1.auto/R/clouderrorreporting_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
ReportedErrorEvent Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An error event which is reported to the Error Reporting system.
}
\seealso{
Other res: \code{\link{DeleteEventsResponse}},
  \code{\link{ErrorContext}}, \code{\link{ErrorEvent}},
  \code{\link{ErrorGroupStats}}, \code{\link{ErrorGroup}},
  \code{\link{HttpRequestContext}},
  \code{\link{ListEventsResponse}},
  \code{\link{ListGroupStatsResponse}},
  \code{\link{ReportErrorEventResponse}},
  \code{\link{ServiceContext}},
  \code{\link{SourceLocation}}, \code{\link{TimedCount}},
  \code{\link{TrackingIssue}}
}

