% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouderrorreporting_objects.R
\name{ReportErrorEventResponse}
\alias{ReportErrorEventResponse}
\title{ReportErrorEventResponse Object}
\usage{
ReportErrorEventResponse()
}
\value{
ReportErrorEventResponse object
}
\description{
ReportErrorEventResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response for reporting an individual error event.Data may be added to this message in the future.
}
