% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouderrorreporting_objects.R
\name{ServiceContext}
\alias{ServiceContext}
\title{ServiceContext Object}
\usage{
ServiceContext(service = NULL, version = NULL)
}
\arguments{
\item{service}{An identifier of the service, such as the name of the}

\item{version}{Represents the source code version that the developer provided,}
}
\value{
ServiceContext object
}
\description{
ServiceContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes a running service that sends errors.Its version changes over time and multiple versions can run in parallel.
}
\seealso{
Other res: \code{\link{DeleteEventsResponse}},
  \code{\link{ErrorContext}}, \code{\link{ErrorEvent}},
  \code{\link{ErrorGroupStats}}, \code{\link{ErrorGroup}},
  \code{\link{HttpRequestContext}},
  \code{\link{ListEventsResponse}},
  \code{\link{ListGroupStatsResponse}},
  \code{\link{ReportErrorEventResponse}},
  \code{\link{ReportedErrorEvent}},
  \code{\link{SourceLocation}}, \code{\link{TimedCount}},
  \code{\link{TrackingIssue}}
}

