% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouderrorreporting_objects.R
\name{SourceLocation}
\alias{SourceLocation}
\title{SourceLocation Object}
\usage{
SourceLocation(filePath = NULL, functionName = NULL, lineNumber = NULL)
}
\arguments{
\item{filePath}{The source code filename, which can include a truncated relative}

\item{functionName}{Human-readable name of a function or method}

\item{lineNumber}{1-based}
}
\value{
SourceLocation object
}
\description{
SourceLocation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Indicates a location in the source code of the service for whicherrors are reported.This data should be provided by the application when reporting an error,unless the error report has been generated automatically from Google AppEngine logs. All fields are optional.
}
\seealso{
Other res: \code{\link{DeleteEventsResponse}},
  \code{\link{ErrorContext}}, \code{\link{ErrorEvent}},
  \code{\link{ErrorGroupStats}}, \code{\link{ErrorGroup}},
  \code{\link{HttpRequestContext}},
  \code{\link{ListEventsResponse}},
  \code{\link{ListGroupStatsResponse}},
  \code{\link{ReportErrorEventResponse}},
  \code{\link{ReportedErrorEvent}},
  \code{\link{ServiceContext}}, \code{\link{TimedCount}},
  \code{\link{TrackingIssue}}
}

