% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouderrorreporting_objects.R
\name{TimedCount}
\alias{TimedCount}
\title{TimedCount Object}
\usage{
TimedCount(startTime = NULL, endTime = NULL, count = NULL)
}
\arguments{
\item{startTime}{Start of the time period to which `count` refers (included)}

\item{endTime}{End of the time period to which `count` refers (excluded)}

\item{count}{Approximate number of occurrences in the given time period}
}
\value{
TimedCount object
}
\description{
TimedCount Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The number of errors in a given time period.All numbers are approximate since the error events are sampledbefore counting them.
}
\seealso{
Other res: \code{\link{DeleteEventsResponse}},
  \code{\link{ErrorContext}}, \code{\link{ErrorEvent}},
  \code{\link{ErrorGroupStats}}, \code{\link{ErrorGroup}},
  \code{\link{HttpRequestContext}},
  \code{\link{ListEventsResponse}},
  \code{\link{ListGroupStatsResponse}},
  \code{\link{ReportErrorEventResponse}},
  \code{\link{ReportedErrorEvent}},
  \code{\link{ServiceContext}},
  \code{\link{SourceLocation}}, \code{\link{TrackingIssue}}
}

