% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouderrorreporting_objects.R
\name{DeleteEventsResponse}
\alias{DeleteEventsResponse}
\title{Stackdriver Error Reporting API Objects 
Stackdriver Error Reporting groups and counts similar errors from cloud services. The Stackdriver Error Reporting API provides a way to report new errors and read access to error groups and their associated errors.}
\usage{
DeleteEventsResponse()
}
\value{
DeleteEventsResponse object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:32:10
filename: /Users/mark/dev/R/autoGoogleAPI/googleclouderrorreportingv1beta1.auto/R/clouderrorreporting_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
DeleteEventsResponse Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for deleting error events.
}
