% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouderrorreporting_objects.R
\name{ErrorContext}
\alias{ErrorContext}
\title{ErrorContext Object}
\usage{
ErrorContext(httpRequest = NULL, user = NULL, reportLocation = NULL)
}
\arguments{
\item{httpRequest}{The HTTP request which was processed when the error was}

\item{user}{The user who caused or was affected by the crash}

\item{reportLocation}{The location in the source code where the decision was made to}
}
\value{
ErrorContext object
}
\description{
ErrorContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A description of the context in which an error occurred.This data should be provided by the application when reporting an error,unless theerror report has been generated automatically from Google App Engine logs.
}
