% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouderrorreporting_objects.R
\name{ErrorEvent}
\alias{ErrorEvent}
\title{ErrorEvent Object}
\usage{
ErrorEvent(serviceContext = NULL, context = NULL, eventTime = NULL,
  message = NULL)
}
\arguments{
\item{serviceContext}{The `ServiceContext` for which this error was reported}

\item{context}{Data about the context in which the error occurred}

\item{eventTime}{Time when the event occurred as provided in the error report}

\item{message}{The stack trace that was reported or logged by the service}
}
\value{
ErrorEvent object
}
\description{
ErrorEvent Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An error event which is returned by the Error Reporting system.
}
\seealso{
Other res: \code{\link{DeleteEventsResponse}},
  \code{\link{ErrorContext}},
  \code{\link{ErrorGroupStats}}, \code{\link{ErrorGroup}},
  \code{\link{HttpRequestContext}},
  \code{\link{ListEventsResponse}},
  \code{\link{ListGroupStatsResponse}},
  \code{\link{ReportErrorEventResponse}},
  \code{\link{ReportedErrorEvent}},
  \code{\link{ServiceContext}},
  \code{\link{SourceLocation}}, \code{\link{TimedCount}},
  \code{\link{TrackingIssue}}
}

