% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouderrorreporting_objects.R
\name{ErrorGroup}
\alias{ErrorGroup}
\title{ErrorGroup Object}
\usage{
ErrorGroup(name = NULL, groupId = NULL, trackingIssues = NULL)
}
\arguments{
\item{name}{The group resource name}

\item{groupId}{Group IDs are unique for a given project}

\item{trackingIssues}{Associated tracking issues}
}
\value{
ErrorGroup object
}
\description{
ErrorGroup Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Description of a group of similar error events.
}
