% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouderrorreporting_objects.R
\name{ErrorGroupStats}
\alias{ErrorGroupStats}
\title{ErrorGroupStats Object}
\usage{
ErrorGroupStats(group = NULL, firstSeenTime = NULL, count = NULL,
  affectedUsersCount = NULL, lastSeenTime = NULL,
  numAffectedServices = NULL, affectedServices = NULL,
  representative = NULL, timedCounts = NULL)
}
\arguments{
\item{group}{Group data that is independent of the filter criteria}

\item{firstSeenTime}{Approximate first occurrence that was ever seen for this group}

\item{count}{Approximate total number of events in the given group that match}

\item{affectedUsersCount}{Approximate number of affected users in the given group that}

\item{lastSeenTime}{Approximate last occurrence that was ever seen for this group and}

\item{numAffectedServices}{The total number of services with a non-zero error count for the given}

\item{affectedServices}{Service contexts with a non-zero error count for the given filter}

\item{representative}{An arbitrary event that is chosen as representative for the whole group}

\item{timedCounts}{Approximate number of occurrences over time}
}
\value{
ErrorGroupStats object
}
\description{
ErrorGroupStats Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Data extracted for a specific group based on certain filter criteria,such as a given time period and/or service filter.
}
