% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouderrorreporting_objects.R
\name{HttpRequestContext}
\alias{HttpRequestContext}
\title{HttpRequestContext Object}
\usage{
HttpRequestContext(method = NULL, remoteIp = NULL, referrer = NULL,
  userAgent = NULL, url = NULL, responseStatusCode = NULL)
}
\arguments{
\item{method}{The type of HTTP request, such as `GET`, `POST`, etc}

\item{remoteIp}{The IP address from which the request originated}

\item{referrer}{The referrer information that is provided with the request}

\item{userAgent}{The user agent information that is provided with the request}

\item{url}{The URL of the request}

\item{responseStatusCode}{The HTTP response status code for the request}
}
\value{
HttpRequestContext object
}
\description{
HttpRequestContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
HTTP request data that is related to a reported error.This data should be provided by the application when reporting an error,unless theerror report has been generated automatically from Google App Engine logs.
}
