% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouderrorreporting_objects.R
\name{ListEventsResponse}
\alias{ListEventsResponse}
\title{ListEventsResponse Object}
\usage{
ListEventsResponse(nextPageToken = NULL, errorEvents = NULL)
}
\arguments{
\item{nextPageToken}{If non-empty, more results are available}

\item{errorEvents}{The error events which match the given request}
}
\value{
ListEventsResponse object
}
\description{
ListEventsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains a set of requested error events.
}
\seealso{
Other res: \code{\link{DeleteEventsResponse}},
  \code{\link{ErrorContext}}, \code{\link{ErrorEvent}},
  \code{\link{ErrorGroupStats}}, \code{\link{ErrorGroup}},
  \code{\link{HttpRequestContext}},
  \code{\link{ListGroupStatsResponse}},
  \code{\link{ReportErrorEventResponse}},
  \code{\link{ReportedErrorEvent}},
  \code{\link{ServiceContext}},
  \code{\link{SourceLocation}}, \code{\link{TimedCount}},
  \code{\link{TrackingIssue}}
}

