% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouderrorreporting_objects.R
\name{ListGroupStatsResponse}
\alias{ListGroupStatsResponse}
\title{ListGroupStatsResponse Object}
\usage{
ListGroupStatsResponse(nextPageToken = NULL, errorGroupStats = NULL)
}
\arguments{
\item{nextPageToken}{If non-empty, more results are available}

\item{errorGroupStats}{The error group stats which match the given request}
}
\value{
ListGroupStatsResponse object
}
\description{
ListGroupStatsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains a set of requested error group stats.
}
\seealso{
Other res: \code{\link{DeleteEventsResponse}},
  \code{\link{ErrorContext}}, \code{\link{ErrorEvent}},
  \code{\link{ErrorGroupStats}}, \code{\link{ErrorGroup}},
  \code{\link{HttpRequestContext}},
  \code{\link{ListEventsResponse}},
  \code{\link{ReportErrorEventResponse}},
  \code{\link{ReportedErrorEvent}},
  \code{\link{ServiceContext}},
  \code{\link{SourceLocation}}, \code{\link{TimedCount}},
  \code{\link{TrackingIssue}}
}

