% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouderrorreporting_objects.R
\name{ReportedErrorEvent}
\alias{ReportedErrorEvent}
\title{ReportedErrorEvent Object}
\usage{
ReportedErrorEvent(context = NULL, message = NULL, serviceContext = NULL,
  eventTime = NULL)
}
\arguments{
\item{context}{[Optional] A description of the context in which the error occurred}

\item{message}{[Required] The error message}

\item{serviceContext}{[Required] The service context in which this error has occurred}

\item{eventTime}{[Optional] Time when the event occurred}
}
\value{
ReportedErrorEvent object
}
\description{
ReportedErrorEvent Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An error event which is reported to the Error Reporting system.
}
