% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouderrorreporting_objects.R
\name{ReportErrorEventResponse}
\alias{ReportErrorEventResponse}
\title{ReportErrorEventResponse Object}
\usage{
ReportErrorEventResponse()
}
\value{
ReportErrorEventResponse object
}
\description{
ReportErrorEventResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response for reporting an individual error event.Data may be added to this message in the future.
}
\seealso{
Other res: \code{\link{DeleteEventsResponse}},
  \code{\link{ErrorContext}}, \code{\link{ErrorEvent}},
  \code{\link{ErrorGroupStats}}, \code{\link{ErrorGroup}},
  \code{\link{HttpRequestContext}},
  \code{\link{ListEventsResponse}},
  \code{\link{ListGroupStatsResponse}},
  \code{\link{ReportedErrorEvent}},
  \code{\link{ServiceContext}},
  \code{\link{SourceLocation}}, \code{\link{TimedCount}},
  \code{\link{TrackingIssue}}
}

