% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouderrorreporting_objects.R
\name{ServiceContext}
\alias{ServiceContext}
\title{ServiceContext Object}
\usage{
ServiceContext(service = NULL, resourceType = NULL, version = NULL)
}
\arguments{
\item{service}{An identifier of the service, such as the name of the}

\item{resourceType}{Type of the MonitoredResource}

\item{version}{Represents the source code version that the developer provided,}
}
\value{
ServiceContext object
}
\description{
ServiceContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes a running service that sends errors.Its version changes over time and multiple versions can run in parallel.
}
