% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouderrorreporting_objects.R
\name{SourceLocation}
\alias{SourceLocation}
\title{SourceLocation Object}
\usage{
SourceLocation(functionName = NULL, filePath = NULL, lineNumber = NULL)
}
\arguments{
\item{functionName}{Human-readable name of a function or method}

\item{filePath}{The source code filename, which can include a truncated relative}

\item{lineNumber}{1-based}
}
\value{
SourceLocation object
}
\description{
SourceLocation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Indicates a location in the source code of the service for whicherrors are reported.This data should be provided by the application when reporting an error,unless the error report has been generated automatically from Google AppEngine logs. All fields are optional.
}
