% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouderrorreporting_objects.R
\name{TimedCount}
\alias{TimedCount}
\title{TimedCount Object}
\usage{
TimedCount(endTime = NULL, count = NULL, startTime = NULL)
}
\arguments{
\item{endTime}{End of the time period to which `count` refers (excluded)}

\item{count}{Approximate number of occurrences in the given time period}

\item{startTime}{Start of the time period to which `count` refers (included)}
}
\value{
TimedCount object
}
\description{
TimedCount Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The number of errors in a given time period.All numbers are approximate since the error events are sampledbefore counting them.
}
